Make teuthology.suite a subpackage
